backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn==0.22.0
pydantic==2.10.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==24.1.0
jinja2==3.1.2
websockets==11.0.3

# Browser automation
selenium==4.15.0
webdriver-manager==4.0.1
playwright==1.49.0

# Data processing
beautifulsoup4==4.12.2
lxml==5.3.0
pillow==10.4.0
numpy==2.2.2

# Utilities
rich==13.9.4
python-dateutil==2.8.2
requests==2.31.0
urllib3==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6

# Cache and storage
aioredis==2.0.1
motor==3.3.2
pymongo==4.6.1

# Security
cryptography==42.0.2
bcrypt==4.1.2
PyJWT==2.8.0

# Compression and serialization
msgpack==1.0.7
ujson==5.9.0
